We were mistakenly working with two objects that were able to handle preferences/properties. Rationalise them
