+++ org.mwc.debrief.lite/src/main/java/org/mwc/debrief/lite/LiteApplication.java
@@ -18,10 +18,43 @@ import java.awt.MenuShortcut;
 
 import Debrief.GUI.Frames.Application;
 import Debrief.GUI.Frames.Session;
+import Debrief.ReaderWriter.Replay.ImportReplay;
+import Debrief.ReaderWriter.Replay.ImportReplay.ProvidesModeSelector;
 import MWC.GUI.Tool;
 
-class LiteApplication extends Application
+class LiteApplication extends Application implements ProvidesModeSelector
 {
+  
+  /** store preferences for import mode, and import frequency
+   * 
+   */
+  private final ImportSettings settings;
+  
+  private final Long freq;
+
+  public LiteApplication(final String mode, final long freq)
+  {
+    settings = new ImportSettings(mode, freq);
+    this.freq = freq;
+  }
+  
+  @Override
+  public String getProperty(final String name)
+  {
+    if (name.equals(ImportReplay.TRACK_IMPORT_MODE))
+    {
+      return settings.importMode;
+    }
+    else if (name.equals(ImportReplay.RESAMPLE_FREQUENCY))
+    {
+      return "" + settings.sampleFrequency;
+    }
+    else
+    {
+      return super.getProperty(name);
+    }
+  }
+  
   @Override
   protected void addMenuItem(final String theMenu, final String theLabel,
       final Tool theTool, final MenuShortcut theShortCut)
@@ -83,4 +116,16 @@ class LiteApplication extends Application
     throw new IllegalArgumentException("Not implemented");
   }
 
+
+  @Override
+  public Long getSelectedImportFrequency(final String trackName)
+  {
+    return this.freq;
+  }
+
+  @Override
+  public ImportSettings getSelectedImportMode(final String trackName)
+  {
+    return settings;
+  }
 }
